pyproject.toml
[project]
name = "ai-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector,openai,deepgram]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-fishaudio~=1.3",
    "python-dotenv>=1.2.1",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "sqlalchemy>=2.0.0",
    "aiomysql>=0.2.0",
]
